requirement.txt
streamlit
mtcnn
opencv-python
numpy
Pillow
